+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- `--use-implicit-cmds`.
+- `--convert-segments`.
+
 ## [0.7.1] - 2016-11-04
 ### Added
 - Check that input file exist.